vinylcutter::inkcut: force inkcut static config
This uses home-manager to force the inkcut config files to be exactly
what we've defined in this repo, no funny business or users modifying
settings. This configuration is static and should always work unti the
hardware changes. When the hardware changes, it requires a re-evaluation
by a maintainer/person anyway.

https://github.com/DoESLiverpool/somebody-should/issues/1208