flake: switch to nixos-21.05
nixos-unstable has too many breaking changes for the software we're running, it's best to wait for NixOS 21.11 before attepmting to do things like upgrade MQTT and Node-RED. MQTT v3.1.1 breaks clients which do not set a keepalive in the packet, whereas Node-RED doesn't use or migrate the old JSON files. It's all a pain.